Use shwrap* helpers from coreos-ci-lib in prod pipelines
This introduces coreos-ci-lib in the production jobs and starts making
use of its shwrap helpers. This allows us to drain the duplicates we
have here in the "utils" module.

There's more stuff we can drain from "utils" into coreos-ci-lib, but
this at least paves the way for using more coreos-ci-lib helpers in
these jobs. (For example, we should switch over to `fcosBuild`,
`fcosKola`, etc... the way it's used in upstream CI). Notably, this will
also allow us to leveraging the gangplank work going on in
coreos-ci-lib.